add staff-configurable announcements (#36)
* initial work on staff-editable announcements

* improve announcements with deletion too

* display global announcements by storing in app config (avoid repeated db reads)

* move announcement types into config

* fixup CSS and markup for announcements

* tidy up announcements: be pessimistic about booleans (disallow HTML)

* add example to db if announcements table is empty

* update announcements after a deletion

(otherwise need to reboot to dismiss it)

closes #11 — including seeding the example announcement message (suspending races) if the announcement table is empty when the app starts up... a bit cheeky but good for now.

The sneaky thing is that this is using Flask's config object as a persistent memory store which might not be sound, not sure if there's something complex going on with sharding and threads in Flask but I am pretty sure there isn't. Let's see eh?