put setup code into create_app instead of using the decorator
it's not firing for me, running via gunicorn, and this seems simpler:
the (coding) problem I was having was lack of the app context for the
ORM and the (simple?) solution is to use with app.app_context() which
is working for me now.

Note this also removes the check-for-None because that was a safety
that was only needed on the first call: now we _know_ this is the
first call (because it's at time of creation).

Could investigate why the Flask decorator isn't working as expected
but I don't feel that's a problem that needs solving here: not
using the decorator is simpler that using it, because we now
know we're always going via app.py's create_app() (nice work there :-))