explicitly force config names to be string names==values
to minimise the risk of typos/etc sneaking in, using Enum's auto() to
force the name to be the same as value. There's still an annoying risk
of using the enum as a key into the config object without explicitly
using .name (or .value, which is now guaranteed to be the same).