fix fake-LaTeX CSS (and make it entirely optional)
This had been broken in the move into Pelican, and wasn't important
enough to get attention at the time. Now it's working, but as it is
silly (and more importantly the CSS toggle in JS is not great;
there's a flash of unstyled content before it changes), it needed
to be made optional too. So there's a new setting in the Tech Notes
config group which switches it off completely.
Because it's silly, it's off by default.

Being able to choose between light and dark mode, however, is
*not* silly, but maybe the right way to do that is with two
builds of static content — and not using this JS toggle which is
modifying the applied style sheet dynamically as the page loads
:-O