v1.0.1 small changes prior to RHUL 2023 start (#111)
* add dev-tool for extracting config setting descs for www docs

maybe remove this later?

* reposition edit user button to be clearer/simpler

* tighten up API handler including documenting it (/api page)

* rename showspecs to show_specs for clarity

* add GNU Affero GPL (was supposed to be there from the start)

Using Affero rather than MIT to encourage exposing source code
to visitors to forked servers (in effect this probably happens
these days simply by having the link the (buggyrace org)
GitHub repo anyway.

* add FOSS notice to the about page

* fix fake-LaTeX CSS (and make it entirely optional)

This had been broken in the move into Pelican, and wasn't important
enough to get attention at the time. Now it's working, but as it is
silly (and more importantly the CSS toggle in JS is not great;
there's a flash of unstyled content before it changes), it needed
to be made optional too. So there's a new setting in the Tech Notes
config group which switches it off completely.
Because it's silly, it's off by default.

Being able to choose between light and dark mode, however, is
*not* silly, but maybe the right way to do that is with two
builds of static content — and not using this JS toggle which is
modifying the applied style sheet dynamically as the page loads
:-O

* bump KeyShapeDiagram JS to cope with Keyshape v1.1 vs 1.2

* add tceh notes Flask webserver text (diagram to be added later)