replace old (jekyll) {{site.thing}} in task list with %THING%
The default tasks.md is still currently very specific to RHUL's
(remote server and VSCode) workflow, need to make this conditional
in some way (tricky as there is no conditional processing in the
markdown creation (yet). But for now it's been tidied up so at
least the subs are correct. (Also removed some explicit links
to videos that have been retained in the local (custom-edited)
one.