v1.0.2 for RHUL (has explicit file tests for static content) (#114)
* add dev-tool for extracting config setting descs for www docs

maybe remove this later?

* reposition edit user button to be clearer/simpler

* tighten up API handler including documenting it (/api page)

* rename showspecs to show_specs for clarity

* add GNU Affero GPL (was supposed to be there from the start)

Using Affero rather than MIT to encourage exposing source code
to visitors to forked servers (in effect this probably happens
these days simply by having the link the (buggyrace org)
GitHub repo anyway.

* add FOSS notice to the about page

* fix fake-LaTeX CSS (and make it entirely optional)

This had been broken in the move into Pelican, and wasn't important
enough to get attention at the time. Now it's working, but as it is
silly (and more importantly the CSS toggle in JS is not great;
there's a flash of unstyled content before it changes), it needed
to be made optional too. So there's a new setting in the Tech Notes
config group which switches it off completely.
Because it's silly, it's off by default.

Being able to choose between light and dark mode, however, is
*not* silly, but maybe the right way to do that is with two
builds of static content — and not using this JS toggle which is
modifying the applied style sheet dynamically as the page loads
:-O

* bump KeyShapeDiagram JS to cope with Keyshape v1.1 vs 1.2

* add tceh notes Flask webserver text (diagram to be added later)

* fix wrong limit on username length (in db it is 80), closes #113

* replace old (jekyll) {{site.thing}} in task list with %THING%

The default tasks.md is still currently very specific to RHUL's
(remote server and VSCode) workflow, need to make this conditional
in some way (tricky as there is no conditional processing in the
markdown creation (yet). But for now it's been tidied up so at
least the subs are correct. (Also removed some explicit links
to videos that have been retained in the local (custom-edited)
one.

* tidy up (and add conditionals) to workspace page

* add web app URL on the remote server, for workspace page

feels like this should be showing up somewhere else too

* show staff users at top of user table, closes #110

* make user table hover on staff much more subtle (it's distracting)

* fix staff/admin display fault on dashboard: all staff are grouped, not just admins

* actively test for presence of static content files for warnings, closes #94

* bump manual version number