v1.0.4 release for RHUL: Flask webserver animation in tech notes, button to remove GitHub details (#131)
* add dev-tool for extracting config setting descs for www docs

maybe remove this later?

* reposition edit user button to be clearer/simpler

* tighten up API handler including documenting it (/api page)

* rename showspecs to show_specs for clarity

* add GNU Affero GPL (was supposed to be there from the start)

Using Affero rather than MIT to encourage exposing source code
to visitors to forked servers (in effect this probably happens
these days simply by having the link the (buggyrace org)
GitHub repo anyway.

* add FOSS notice to the about page

* fix fake-LaTeX CSS (and make it entirely optional)

This had been broken in the move into Pelican, and wasn't important
enough to get attention at the time. Now it's working, but as it is
silly (and more importantly the CSS toggle in JS is not great;
there's a flash of unstyled content before it changes), it needed
to be made optional too. So there's a new setting in the Tech Notes
config group which switches it off completely.
Because it's silly, it's off by default.

Being able to choose between light and dark mode, however, is
*not* silly, but maybe the right way to do that is with two
builds of static content — and not using this JS toggle which is
modifying the applied style sheet dynamically as the page loads
:-O

* bump KeyShapeDiagram JS to cope with Keyshape v1.1 vs 1.2

* add tceh notes Flask webserver text (diagram to be added later)

* fix wrong limit on username length (in db it is 80), closes #113

* replace old (jekyll) {{site.thing}} in task list with %THING%

The default tasks.md is still currently very specific to RHUL's
(remote server and VSCode) workflow, need to make this conditional
in some way (tricky as there is no conditional processing in the
markdown creation (yet). But for now it's been tidied up so at
least the subs are correct. (Also removed some explicit links
to videos that have been retained in the local (custom-edited)
one.

* tidy up (and add conditionals) to workspace page

* add web app URL on the remote server, for workspace page

feels like this should be showing up somewhere else too

* show staff users at top of user table, closes #110

* make user table hover on staff much more subtle (it's distracting)

* fix staff/admin display fault on dashboard: all staff are grouped, not just admins

* actively test for presence of static content files for warnings, closes #94

* bump manual version number

* bring readme/dev notes more up to date

* add day-of-week to submission date display, closes #117

* move staff users to bottom of user form

* bump session activity timestamp for logged-in users

* update labels on last logged-in (now activity) timestamp

* force UTC on timestamps on way into db by explicit UTC in now()

* handle timezone in Pelican dictionary (must be quoted)

* make Europe/London the default timezone, anticipating UK deployment

* for convenience, redirect (incorrect) publish GET requests

handy when refreshing page after testing it

* add servertime_str for server timezone: does not apply daylight saving

it should but too many hours burned on fighting Python's cleverness, will come back to it

* use servertime for localising presentation of tech notes publication date

* remove debug info from API error (it's OK, hasn't been enabled in prod yet)

* add white background to (JS) buggy button in users list for clarity

* move buggy-editor admin button to more sensible location (config)

* update auth code explanation to be more useful/explanatory

* add new get-editor announcement for GET-0 instructions

* bump patch level up for RHUL pre-lab release

* bugfix: delete form only possible if announcement exists

* add delete (user's) GitHub details capability for admin, closes #123

* bugfix: display email on user page (not last name again), closes #126

* bugfix: use username (not example) if available for zipfile, closes #125

* simplify the http status 400 error message

* use Jinja filter to convert UTC times when passing out to front-end

Closes #116 and #127

* bugfix: fix incorrect use of CSV reader, was swallowing first line, oops

* bugfix: handle seconds (strpformat elements aren't optional, oops) servertime

* add link to Flask-as-webserver tech note from index

* add webserver flask animated diagram

* fix some typos in webserver text (minor CSS in diagram too)

* bump the manual version for next RHUL release

* fix user sorted: staff users are shown _last_