fix buggy URL csv migration by dropping the column
The UNIQUE constraint was causing problems for a column that
isn't being used anyway, and the flask-generated migration
turned out to be mySQL-specific (my dev db) so to make it
simpler, just dropping the column is easier as that _does_
work in a database-dialect agnostic way. Note next commit
fixes the code that expects that column to be present,
but this commit is in isolation for testing on (Postgres)
heroku staging server.