add full editor source code to server repo
This could have been done with git submodules, but the risk fo the
files being missed because submodules are not loaded by default, and
the mechanism is not so widely understood by non-expert git users
was too big for the complexity. It's easier to keep it simple and
manual (we don't anticipate the editor changing very much).

This now allows the _default_ behaviour on setup to *not* use
GitHub at all, but to offer a customised ready-to-run buggy editor
as a downloadable zip from this race server.

If your want to encourage your students to use GitHub then you
can enable that functionality (including of course the automatic
forking and injection of issues that was the oringial feature).