do somersaults to pass db URI password in as password query param
Not clear why this is needed by since upgrading Flask, the pw in
the DB URI is not being passed to flask's migrations unless it's
as a password query param, so this rewrites it using the regex.
Seems... unlikely to really be necessary but it's what worked
locally.
Might turn out there's a dependency missing in the currently
updated Flask which is responsible for this silliness.