fix (maybe) old-(?) server URL from config when loading racetrack
First time seen this _not_ on the default localhost:8000 server
but the config setting in the static method was not using the
most recent setting. Heads up this might be a warning shot
re: any config settings being accessed in static methods?
This was via current_app.config[key] so... unexpected that this
problem arose. Testing on remote (Heroku) server with this
commit.