make autogeneration of static content workable, at last
Now the autogeneration of static content is not too onerous even
if it happens on frequent gunicorn worker timeout (which was a small
worry). It only regenerates content if

(a) there's a timestamp when this was done before
(b) the file is missing (i.e., os.path.exists check fails)

This means we're not regenerating content that's already there
which was the lazy case before.

Outstanding problem: can't do this for the zip file without
saving the customised README in the database, argh (and also
the value of is_updating_app_py when it was done), gah