v2.0.1 changes (#188)
* fix (maybe) old-(?) server URL from config when loading racetrack

First time seen this _not_ on the default localhost:8000 server
but the config setting in the static method was not using the
most recent setting. Heads up this might be a warning shot
re: any config settings being accessed in static methods?
This was via current_app.config[key] so... unexpected that this
problem arose. Testing on remote (Heroku) server with this
commit.

* roll back ccd20fa, turns out hadn't saved the setting oops

* fix checkbox is_active input by using yes/no dropdown

* change default: apply titlecase to usernames by default

* bump editor source (security bump on requirements)

* bugfix: hide total cell on users page if not using github, closes #186

* small layout tidy on user's buggy pages

* remove not-connected-to-github warning if not using github

* move jinja filter declarations so they're in place before static regeneration

* make autogeneration of static content workable, at last

Now the autogeneration of static content is not too onerous even
if it happens on frequent gunicorn worker timeout (which was a small
worry). It only regenerates content if

(a) there's a timestamp when this was done before
(b) the file is missing (i.e., os.path.exists check fails)

This means we're not regenerating content that's already there
which was the lazy case before.

Outstanding problem: can't do this for the zip file without
saving the customised README in the database, argh (and also
the value of is_updating_app_py when it was done), gah

* remove debug output from race upload code, oops

* tidy race run: exit if race file not found, don't do ugly sulk-crash

* bump version number to v2.0.1+dev while testing

* refactor generation of editor zipfile so autogenerate works better/at all

* update config description for autogenerating static content which now uh works

* tidy up horrid dependencies in node, thanks to Tom

* tidy up env.example (Docker prefering .env over 'environment')

* update package.json with flask-server-dev

* change default setting for student using GitHub must be 0

otherwise is_using_github defaults to true on the home page

* bump config (and schema.sql) for v2.0.1 release

---------

Co-authored-by: Tom Kuipers <tom@kprs.me>