make issues CSV header configurable but with useful default
The default (title, description) doesn't break existing behaviour
but does support GitLab (which we're using in RHUL t3 this year),
which expects a header row of "title,description". Internally, the
issue parser the server uses refers to these as "title" and "body"
but that is internal — the heading setting doesn't matter there as
it's really only about what is shown at the top of the produced
CSV file.