add 405 (Wrong Method) error handler
Flask doesn't add an Allow header to its response — which it should —
so we're adding it explicitly here (the exception helpfully *does*
have a valid_methods array we can use). But needed to explicitly
create a response here in order to manually add the header, instead
of just returning the rendered page.