add "..." button to admin menu for "manage/admin" pages
The "..." button uses Javascript to slide the admin menu down, but if there's
no javascript, the extra menu is _always_ available from the dashboard.

Decided to show the admin buttons even if account is TA and doesn't have
access, so that the interface/menus are the same for all staff — even though
currently this means the extra menu is mostly pointless for TAs (currently
the only page they can go to from there is the API spec. Need to check if
TAs can allocated API keys (if not... is this a config we need?)