admin changes for v2.0.5 (#218)
* Issue 212 error 405 (#214)

* add 405 (Wrong Method) error handler

Flask doesn't add an Allow header to its response — which it should —
so we're adding it explicitly here (the exception helpfully *does*
have a valid_methods array we can use). But needed to explicitly
create a response here in order to manually add the header, instead
of just returning the rendered page.

* update API specification: fix typos and point out TLS/https

* add toc-links to API specification page

* change system info page title to clarify these are different from config settings

* add "..." button to admin menu for "manage/admin" pages

The "..." button uses Javascript to slide the admin menu down, but if there's
no javascript, the extra menu is _always_ available from the dashboard.

Decided to show the admin buttons even if account is TA and doesn't have
access, so that the interface/menus are the same for all staff — even though
currently this means the extra menu is mostly pointless for TAs (currently
the only page they can go to from there is the API spec. Need to check if
TAs can allocated API keys (if not... is this a config we need?)

* make date-to-string a little more robust (returns None if bad date)

* add age-in-seconds filter for dates in Jinja templates

really for alerting when task notes need publishing

* add timestamp of when tasks were edited to config

This is useful to alerting (in interface) the need to re-publish
the task list. Could have re-used the existing timestamp of when
the tasks were loaded, but it might be handy in diagnostic circumstances
to know the difference

* tidy task-list edit/create timestamps for red buttons

* use Jinja namespace to set the task/tech-notes button-colouring settings

can't usefully push them to the templates as variables because it's used
in the admin nav on every admin page

* make the ... manage/extra admin nav a toggle for hiding too

* make TAs access to setting API keys configurable (default: enabled) #216

* small improvements to navigation for staff accounts using API test

* improve text on docs helper page for clarity to staff

* add whitespace for clarity

* add duplicate change password button to user page (cos it's probably most useful)

* add _IS_DOCS_HELPER_PAGE_ENABLED config to suppress doc helper except when really needed

* add better (more precise) link to example of docs helper script output

* bugfix: missing s from https in description, oops

* update CSS to prevent keyword/json-names word-wrapping in tabulated data

* allow all staff users access to Setup summary

This is because it's now part of the top-level admin nav, much like
the dashboard. There might be links/buttons off the summary that
TAs can't access, but that's fine (maybe disable those buttons, if
it's confusing otherwise?).

* add comment icon to user comments

the multi-spoked asterisk was too obscure!
This adds the "bubble" icon from the Icomoon set — see the
DEVELOPMENT.md document for (updated) details on how to add
more icons like this to the site, if needed.

* use bubble/comment icon on users' adim page

(late commit: should have been part of 6a03607)

* change page title: buggy editor is not necessarily a repo

* rearrange Jinja comment whitespace replacement

moving it out from the class declaration because the nested
double-quotes were breaking VS Code's syntax highlighter because
escaped-quotes-in-quotes were a bit too horrible

* include 'Auth code:' and well as 'Authorisation code:' in special error CSS class

* add link to download button to help/clarify how to save before delete

* show disabled download buttons (and explain them) rather than hiding them on task text page

* tidy display of text task buttons

* use pre to display user comment on user page (consistent with other places now)

* delete task texts (with warnings/confirm) when loading tasks: closes #213

* bump in-source version to v2.0.5