collapse all config setting forms on settings index page
it's confusing and risky (because there are lots of forms on the
single page) and it's nearly always better to go to the group
itself. So now this is concealed behind some javascript for either
jumping to the group you really need by picking the setting you
want, or (reluctantly) letting you reveal all the forms like it
was before (this does have utility: can search the whole page with
command-F).

Degrades gracefully without JS (same as before).

Note there was a problem with Firefox not firing "change" events
on the text input, but it turns out FF is right and the other
browsers a little too trigger-happy with the event: now using
"input" event instead.