add teardown_request to explicitly rollback db transactions
It looks like the timeout blocking/500 problems we're having are caused
by database transactions not being rolled back, and maybe this is it:
there's no _explicit_ rollback in the event of an error.

Haven't identified a single cause of when these are happening but the
heroku logs include suggestion to rollback, so let's see if this fixes
it.