add text-author admin summary, #239
Helps staff keep track of who has written texts and who hasn't,
presented in more convenient way than the (also useful, but
different) matrix.

There probably will be more investigation tools to add to this
page (including inspecting texts and their dates (the dates are
already on this page, but not yet presented to the user).

Note this has also changed the URL so (unlikely but...) old
bookmarks to the texts matrix will need updating.