show warning in admin if race has URLs with protocol that differs from server
This is because it took me too long to notice that my local dev site was
trying to load https: racetrack image because I had carelessly edited the
imported data value. At least there's a change it would be noticed sooner
this way (maybe should be adding this as a warning after _any_ race edit
save.