Issue 277 task list v2.0.76 (#278)
* move task list publication into the database because of problems with ephemeral files

It looks like the Heroku restarts (e.g., just changing an ENV var to trigger it)
may be using an older (persistent? surely not) version of the file when auto-
generating the task list. That seems unlikely but can't reproduce locally, so
solution is simply to switch to storing it in the database anyway. A db read
is still a better hit than rendering the (markdown) task HTML from all the
task records every hit, so publishing still makes sense.

* add note if task-list is stored in DB, with link to config

* bump config to v2.0.76 (task list stored in db)