add ghosted login link to setup nav (closes #281)
This is a bit dirty, but by providing a non-attention-grabbing
login link, we're protected against setup interruption. See the
issue text (routine setup doesn't encounter this problem, but
if the session set-up status is interfered with there's no naive
way of moving off the screen. It's very much a minority problem
and not really a bug that needs fixing. The concern is if the
login link present is a bigger distraction than the problem of
not having it when it's needed.