add VCS links to user pages, ready for RHUL GitLab
this year we have a (useful but) generic link to student GitLab
because we are not storing VCS username info on the race server,
so make sure the link is being provided anyway. Wording on the
link ("repo" or even "buggy repo") is technically too specific,
but currently the code doesn't have a way of determining if
the URL is to the specific repo. Don't think this is a problem
that needs solving since (at least for our students this year)
there probably won't be any other repos in their GitLab account
anyway...