PR implement "preload" distribution method (issue 284) (#289)
* wip

* swap order of VCS config settings (IS_USING_VCS should be first maybe)

* add bare-bones dev commands to run simplest race server to DEVELOPMENT

* add individual student's repo URLs

* change admin users page to show VCS button using dynamic URL to repo

for RHUL this means not linking direct to the repo

* bump this to v3 because some breaking config changes

* add VCS links to user pages, ready for RHUL GitLab

this year we have a (useful but) generic link to student GitLab
because we are not storing VCS username info on the race server,
so make sure the link is being provided anyway. Wording on the
link ("repo" or even "buggy repo") is technically too specific,
but currently the code doesn't have a way of determining if
the URL is to the specific repo. Don't think this is a problem
that needs solving since (at least for our students this year)
there probably won't be any other repos in their GitLab account
anyway...