add work for v3.0.0 release, for RHUL 2024–25 (#291)
* bugfix: add check for staff a/c when showing race results, closes #253

* tighten up handling staff access/visibiluty of race results

includes making presence of results_uploaded_at useful for showing
"no results" in case of abandoned races: relates to #253 and #129

* bump version number to v2.0.75

* add 'race abandoned' for better no-results handling, see #129

* don't show 'is abandoned' if race results are not visible yet

* tidy indentation in source

* don't serve race file if race was abandoned

Currently there's no provision for indicating empty results inside a race file

* add races.json download button (for 5-RACELOG): #275

* add race adandon page (WiP)

* show warning in admin if race has URLs with protocol that differs from server

This is because it took me too long to notice that my local dev site was
trying to load https: racetrack image because I had carelessly edited the
imported data value. At least there's a change it would be noticed sooner
this way (maybe should be adding this as a warning after _any_ race edit
save.

* implemented abandon race functionality (for 'no-results' races), closes #129

* don't allow replay of abandoned races

* fix typo in form label, see #200

* use raw strings to quote regexps with \s in

* bump runtime for Heroku up to Python 3.12

* bump up required gevent version (dependabot notice)

https://github.com/buggyrace/buggy-race-server/security/dependabot/38

* bump up version of Flask as high as can go with Werkzeug...
Dependabot wants Werkzeug 3.0.3 but latest Flask doesn't support that
so for now, highest version of Flask 2 is probably a step in the the
right direction. Wekzug vuln doesn't look like a problem for running
the race server, since it is an attack via the debugger:
https://github.com/buggyrace/buggy-race-server/security/dependabot/47
https://github.com/buggyrace/buggy-race-server/security/dependabot/46

Highest Wekzeug with the encode_url and decode_url methods is 2.3.6

* explictly use Jinja>=3.1.4 for depedabot alert

https://github.com/buggyrace/buggy-race-server/security/dependabot/48

* add editor host and port config settings, applied to (local) editor source

* updated 'static' editor README to use host/port config settings

* update static editor in-repo files to use new hos/port config

* provide editor host/port config for updating static editor in-repo files

* bump config version to 'dev'

* add EDITOR_DISTRIBUTION_METHOD as new setting

* bugfix/patch admin dashboard task list warning if storing in DB, v2.0.77

* add a wildcard button to the config group buttons

* remove debug output from distrib method chooser

* add suggested config settings by editor distribution method

* move Editor config group before Project

* add missing description for Editor config group

* during setup, override default values to match editor distrib method

* when reporting config diff, check for non-empty settings

if the suggested config setting is NONEMPTY_VALUE, we only check
for it being empty, not for its value being different

* add IS_SHOWING_CONFIG_WARNINGS setting so annoying suggestions can be ignored

* Upgrading to heroku-24 stack

* add `flask run` note to DEVELOPMENT.md

* add ghosted login link to setup nav (closes #281)

This is a bit dirty, but by providing a non-attention-grabbing
login link, we're protected against setup interruption. See the
issue text (routine setup doesn't encounter this problem, but
if the session set-up status is interfered with there's no naive
way of moving off the screen. It's very much a minority problem
and not really a bug that needs fixing. The concern is if the
login link present is a bigger distraction than the problem of
not having it when it's needed.

* Issue 282 git name change (#285)

* sort config names in source code for clarity

* add new VCS_NAME config setting

using abbreviated (VCS) form for brevity when using it in texts

* replace explicit refs to GitHub with CVS_NAME config setting

* update descriptions for VCS-related config, somewhat less GitHub-centric

* change config group name from GITHUB to VCS

* rename BUGGY_EDITOR_GITHUB_URL to BUGGY_EDITOR_REPO_URL

* rename IS_STUDENT_USING_GITHUB_REPO to IS_STUDENT_USING_REPO (as it may not be GitHub)

* add github_username to editable fields for user

* add new config STUDENT_EDITOR_REPO_DOMAIN for building link to students' repos

* add CSS to style setting suggestion recommendations more clearly

* update Editor Distrib methods to include VCS_NAME in suggested settings

* PR implement "preload" distribution method (issue 284) (#289)

* wip

* swap order of VCS config settings (IS_USING_VCS should be first maybe)

* add bare-bones dev commands to run simplest race server to DEVELOPMENT

* add individual student's repo URLs

* change admin users page to show VCS button using dynamic URL to repo

for RHUL this means not linking direct to the repo

* bump this to v3 because some breaking config changes

* add VCS links to user pages, ready for RHUL GitLab

this year we have a (useful but) generic link to student GitLab
because we are not storing VCS username info on the race server,
so make sure the link is being provided anyway. Wording on the
link ("repo" or even "buggy repo") is technically too specific,
but currently the code doesn't have a way of determining if
the URL is to the specific repo. Don't think this is a problem
that needs solving since (at least for our students this year)
there probably won't be any other repos in their GitLab account
anyway...

* move editor download/zip config settings into Editor group

was a bit pokey how they'd ended up in the VCS (formerly GitHub)
group, so good to fix it.
Also remove superflous "BUGGY_" prefix since the other EDITOR
config settings don't use it.

* change site setting names (indexed from 1)

* allow order of Issues CSV to be reversed, #287

* name of social group has changed

* tidy wording of config description for clarity

* Change report poster config ready for RHUL 2024–25 (#290)

* add note and link to admin task page re: CSV issue order

* write newlines (obeying config setting re: CRLF why not?) into CSV

* change report and poster settings to allow more variation

Can now require a report and/or a poster, and have them in-editor
or as documents, and even have the poster as part of the report.

So much trouble caused by a ill-thought out marking scheme by
a non-CompSci department, gah!

* remove spurious closing div on settings page

* remove debug output

* tidy up/order group config settings including new remote one

* add useful link to admin for docs helper

* exclude current announcements from config dumps

* update docs helper page to be more helpful (for me)

* bump version to 3.0.0 (breaking change because of config changes)

Main changes ready for RHUL 2024–5 (term 3) CS1999 module.
Config settings have changed in a few places such that existing
config will not be carried over (including the new separation of
report and poster, and renaming/renumbering "SOCIAL_" external
links to "SITE_" because... they were badly named.
Most signigicant is probably new EDITOR_DISTRIBUTION_METHOD
config setting which we had been trying to infer from all the
other settings. Now it's driven by this one.

Addresses #253 race visibility bugfix with 1944ad1
Closes #275 by providing races.json endpoint (50d386f)
Closes #279 new EDITOR_DISTRIBUTION_METHOD config setting
Closes #282 generalise VCS name (supporting our use of GitLab this year)
Closes #283 exposes VCS username to "normal" admin in backend
(because previously was tightly coupled to the autofork method)
Closes #284 "preload" distrib method for RHUL 2024–25 GitLab use
Closes #286 fix newlines in issues.csv (was actively suppressing them)
Closes #287 allowing issues CSV to be exported in reverse order