replace runtime.txt with .python-version (3.12) closes #293
This is a requirement from Heroku (deprecating use of
runtime.txt. By only specifying the major version, it
allows the latest release of that version to be used.

Not moving to _latest (major) Python version_ yet:
end-of-life is 2028 so got three years oof security
updates yet.
Sticking with 3.12 for now as it's still supported, and
just in case there is anything in the (bootstrap flask)
dependencies that I'm unaware of.

https://devcenter.heroku.com/articles/python-runtimes