force max-width of narrow content, closes #306
There was something pokey happening only on the narrowest
screens (presumably that's Bootstrap's --breakpoint-sm: 576px;)
that was making pre tags's 100% width comfortable with being
wider than viewport, which wasn't happening on wider screens,
despite the pre tag being narrower in those layouts. So this
forces the main container block (which is wrapped by a div
in `layout.html` but previously wasn't styled) to be no
wider than the viewport on very narrow screens. Seems to do
the trick.

Maybe... need to upgrade to a newer version of bootstrap
than that which the original flask-cookiecutter site was
built with.