add (JS-only) show/hide password toggle on passwords/auth codes, closes #304
This isn't implementing a "confirm" option but having a reveal seems like this
is less important.

Have rolled it out to all (explicitly-identified) password and auth-code
inputs, so in general this should be less of a problem.

Progressive enhancement: if JS is available, then the option to reveal/toggle
the password input is provided; otherwise, it's just a `type=password` input
as it was before.