fixing problems with Zip file
Specifically, the EDITOR_DISTRIBUTION_METHOD changes were a bit reckless
concerning the "zip" file handling, and this wasn't spotted because
we (RHUL) aren't using it — we've always distributed via Git one way
or another. But it became clear when tidying up the editor code in
buggy-race-editor repo that it was a little bit broken.
A deceptively minor issue with the localhost rewriting (in the
zip file's app.py) was the trigger: see #294