change default of EDITOR_HOST from 0.0.0.0 to localhost
Now that the default behaviour of IS_WRITING_HOST_IN_EDITOR is "no",
the EDITOR_HOST can be used throuhgout the pages/tech notes as the
host that students use to access it. It's 0.0.0.0 in the editor's
app.py (in the server's copy) and the key thing now is that that
can still remain even while setting this value to someting else.