tidy up (potentially confusing) 'Buggy editor' admin page
For projects which aren't using the internal zip file this page is never
needed and perhaps a bit confusing. So similarly to how the tech notes
admin page is "hidden" if tech notes are not being used, the buggy editor
page will not be linked on the admin nav if the editor distribution
method that's been chosen doesn't use it.

To do this, needed to add a "global function" to the jinja template
since really we're checking the config setting against a range of
values.