improve handling of published task list: closes #277
The bug was causing the task list not not be generated cleanly on
restart, but it turned out it wasn't very clear when the task list
was or was not being stored in the database (DbFiles). So now we
explicitly purge the unwanted (either database record or hard file)
task list to avoid confusion. The presence of a timestamp (in the
config _TASK_LIST_GENERATED_DATETIME) is used to determine if the
file should be there... but only if it is a "static content" file...
if it's in the database there's no need to regenerate it.