remove placeholders from HTML text inputs, closes #323
Kept one in — it's the one on the login form (the default username is
"hamster") — because explicitly putting an example in there made the form
look unlike a login form. Furthermore, that placeholder is configurable
via USERNAME_EXAMPLE which explicitly suggests it can be blank. The
important information being conveyed here is that it's _not_ a college
username (or email address), but that won't always be relevant. This form
is also supported with a page-specific announcement, and, unlike in the
other forms, it's much less likely that "hamster" will be interpreted as
input that's already been made.