add explicit min and max values for data validation in Forms (for #335)
Underlying problem was the input form (through WTF Forms) preventing us
editing (to enable single login) a user with a username of only 2 chars.
Turns out that there were a bunch of hard-coded limits from the
cookiecutter-flask initial user forms.

This fixes this generally, and applies the `MIN_USERNAME_LENGTH` that
was already in `config.py` to the Form validation. And then adds a
whole bunch more of these values.

Now need to check that none of these contradict what we're allowing
in the database/models, just in case that is happening.