provide link to download task markdown (closes #336)
Don't want to put the two "download tasks as Markdown" buttons on
the main tasks admin because it's really about bulk edit/loading
(even though in this case we noticed how hard it was to find when
trying to save them... for bulk loading *next year*).

So the button on the tasks admin links to the (two) download buttons
on the load/bulk-edit page, which needed a small change (heading
and id-for-target) too.

Updated docs to match.